backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-aiplatform==1.52.0
google-cloud-firestore==2.16.0
google-cloud-secret-manager==2.20.0
PyGithub==2.3.0
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
